Fix compilation on MSVC 2010 with the C++0x version of std::bitset
Hopefuly it doesn't break anything on older compilers...


git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/kdesupport/taglib@1216365 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ taglib/ape/apefooter.cpp
@@ -189,7 +189,7 @@ void Footer::parse(const ByteVector &data)
 
   // Read the flags
 
-  std::bitset<32> flags(data.mid(20, 4).toUInt(false));
+  std::bitset<32> flags(static_cast<unsigned long long>(data.mid(20, 4).toUInt(false)));
 
   d->headerPresent = flags[31];
   d->footerPresent = !flags[30];
